refactors context
